+++ example/lib/main.dart
@@ -21,16 +21,15 @@ class _MyAppState extends State<MyApp> {
 
     nImagePickerController
     ..fileTypes = const [ 'png', 'jpg', 'jpeg' ]
-    ..imageKey  = 'upload_image_name';
-
-    nImagePickerController.addListener(() => setState(() {}));
+    ..addListener(() => setState(() {}));
   }
 
   @override
   void dispose() {
     super.dispose();
-    nImagePickerController.removeListener((){});
-    nImagePickerController.dispose();
+    nImagePickerController
+    ..removeListener((){})
+    ..dispose();
   }
 
   @override
@@ -46,7 +45,7 @@ class _MyAppState extends State<MyApp> {
             NImagePicker(
               controller        : nImagePickerController,
               // this is a protected server image, you must to provide a different header in web
-              onLoadingImage    : "https://api.9780bitcoin.com/media/202403/images/company/PHOTO_93080414-8207-490c-ab40-1b4eac7a1076.png",
+              onLoadingImage    : 'https://w.wallhaven.cc/full/49/wallhaven-49d5y8.jpg',
               bankgroundColor   : const Color(0xFFededed).withOpacity(0.8),
               height            : 250,
               width             : 250,
@@ -69,9 +68,8 @@ class _MyAppState extends State<MyApp> {
               alignment: Alignment.center,
               child   :
               InkWell(
-                onTap: () async => await nImagePickerController.multipartFile.then((value) {
-                  debugPrint(value.length.toString());
-                  debugPrint(nImagePickerController.hasNoImage.toString());
+                onTap: () async => await nImagePickerController.image(key: "json_image_key_name").then((value) {
+                  debugPrint(value.filename);
                 }),
                 child: const Text("TEST", style: TextStyle(color: Colors.black)),
               )