+++ lib/src/platform/router.dart
@@ -1,0 +1,3 @@
+import 'package:n_image_picker/src/platform_tools.dart';
+
+PlatformTools getInstance() => throw UnsupportedError('Unsupported Platform');