+++ lib/src/router_helper.dart
@@ -1,3 +1,0 @@
-import 'custom_file.dart';
-
-CustomFile getInstance() => throw UnsupportedError('Unsupported Platform');